[Feat & Docs] : 인터페이스 삭제 및 그에 따른Docs 문서 변경
인터페이스를 통해 추상화 레벨은 증가하였으나, 전체적은 클래스 간의
의존성 관리의 어려움이 발생해 인터페이스를 삭제하였습니다.

전체 프로그램 구현 완료 후 리팩토링 예정입니다.
