--- src/main/java/baseball/BaseballGame.java
@@ -1,5 +1,0 @@
-package baseball;
-
-public class BaseballGame {
-
-}
