--- src/main/java/baseball/BaseballGameApplication.java
@@ -1,5 +1,0 @@
-package baseball;
-
-public class BaseballGameApplication {
-
-}
