--- src/main/java/baseball/Game.java
@@ -1,12 +1,0 @@
-package baseball;
-
-/**
- * 한 주기의 게임 운영을 담당하는 인터페이스
- */
-public interface Game {
-
-    /**
-     * 게임을 실행하는 메서드
-     */
-    void run();
-}
