--- src/main/java/baseball/GameConfiguration.java
@@ -1,8 +1,0 @@
-package baseball;
-
-public interface GameConfiguration {
-    SingleDigitForGame getSingleDigit(int number);
-    MultiDigitNumberForGame getMultiDigit(int number);
-    Game getGame();
-    IOManager getInputReader();
-}
