--- src/main/java/baseball/GameMode.java
@@ -1,2 +1,0 @@
-package baseball;public class GameMode {
-}
