--- src/main/java/baseball/InputReader.java
@@ -1,5 +1,0 @@
-package baseball;
-
-public class InputReader {
-
-}
