+++ src/main/java/baseball/MultiDigitNumber.java
@@ -1,0 +1,26 @@
+package baseball;
+
+/**
+ * 여러개의 SingleDigit 으로 이루어진 MultiDigitNumber 수를 담는 일급 컬랙션 클래스
+ *
+ * @author 호두
+ * @version 1.0
+ */
+public class MultiDigitNumber {
+
+    /**
+     * @param number 입력으로 주어지는 MultiDigitNumber
+     * @return 주어진 number 가 유효한 경우 true, 그렇지 않으면 false 반환
+     */
+    boolean validateMultiDigitNumber(int number){
+        return false;
+    }
+
+    /**
+     * @param otherNumberForGame
+     * @return 두 MultiDigitNumber 간의 비교 결과를 CompareResult 객체에 담아 반환
+     */
+    NumberCompareResult getCompareResult(MultiDigitNumber otherNumberForGame){
+        return null;
+    }
+}
