+++ src/main/java/baseball/NumberBaseballGame.java
@@ -1,0 +1,14 @@
+package baseball;
+
+/**
+ * 한 주기의 게임 운영을 담당하는 클래스
+ */
+public class NumberBaseballGame {
+
+    /**
+     * 게임을 실행하는 메서드
+     */
+    void run(){
+
+    }
+}
