+++ src/main/java/baseball/NumberBaseballGameApplication.java
@@ -1,11 +1,13 @@
 package baseball;
 /**
- * Game 전체 주기를 관리하는 인터페이스.
+ * Game 전체 주기를 관리하는 클래스.
  * 게임의 재실행 및 종료와 같은 기능을 담당
  */
-public interface GameApplication {
+public class NumberBaseballGameApplication {
     /**
      * GameApplication 실행 메서드
      */
-    void execute();
+    void execute(){
+
+    }
 }
