--- src/main/java/baseball/NumberForBaseballGame.java
@@ -1,5 +1,0 @@
-package baseball;
-
-public class NumberForBaseballGame {
-
-}
