--- src/main/java/baseball/RandomNumberGenerator.java
@@ -1,5 +1,0 @@
-package baseball;
-
-public class RandomNumberGenerator {
-
-}
